+++ app/src/main/java/com/shentu/paper/model/body/DelCollectBody.kt
@@ -1,3 +1,3 @@
 package com.shentu.paper.model.body
 
-class DelCollectBody(private val ids: List<Int>)
+class DelCollectBody(private val ids: List<Long>)
